Extract moveGoldens
